backend/requirements.txt
quart
quart-rate-limiter
aiosqlite
ruff
pytest
pytest-asyncio
anyio
itsdangerous
Pillow
aiofiles
